requirements/base.txt
# Python
# ------------------------------------------------------------------------------
pytz==2022.1  # https://github.com/stub42/pytz
python-slugify==6.1.1  # https://github.com/un33k/python-slugify
Pillow==9.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
redis==4.2.2  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py
twilio==7.8.1
timeago==1.0.15
psutil==5.9.0
bcrypt==3.2.0
dacite==1.6.0
awsebcli==3.20.3


# Django
# ------------------------------------------------------------------------------
django==3.1.14  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-allauth==0.50.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==5.2.0  # https://github.com/jazzband/django-redis
django-compressor==2.4  # https://github.com/django-compressor/django-compressor
django-health-check==3.16.4
django-crontab==0.7.1
django-imagekit==4.1.0
django-seed==0.3.1


# Django Settings
# ------------------------------------------------------------------------------
django-cors-headers==3.11.0 # https://github.com/adamchainz/django-cors-headers
drf-pretty-exception-handler==0.1.3
django-health-check==3.16.4


# Django Admin
# ------------------------------------------------------------------------------
https://github.com/leoyunhyung/django-jet/archive/dev.zip
django-reorder-admin==0.3.1
django-admin-relation-links==0.2.5
django-nested-admin==3.4.0
django-import-export==2.7.1
django-inline-actions==2.4.0
django-admin-rangefilter==0.8.4
django-advanced-filters==1.4.0
django-reverse-admin==2.9.6
django-nested-inline==0.4.5


# Django Models
# ------------------------------------------------------------------------------
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-phonenumber-field==6.1.0
phonenumberslite==8.12.46
django-annoying==0.10.6


# Django Views
# ------------------------------------------------------------------------------
django-filter==2.4.0
django-url-filter==0.3.15
django-extra-fields==3.0.2
drf-yasg==1.20.0


# Django REST Framework
# ------------------------------------------------------------------------------
djangorestframework==3.13.1  # https://github.com/encode/django-rest-framework
drf-nested-routers==0.93.4
